--- jms/org.wso2.carbon.transport.jms/src/main/java/org/wso2/carbon/transport/jms/clientfactory/JMSErrorListener.java
@@ -1,41 +1,0 @@
-/*
- *  Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *  WSO2 Inc. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.carbon.transport.jms.clientfactory;
-
-import org.wso2.carbon.transport.jms.factory.JMSImprovedConnectionFactory;
-
-import javax.jms.ExceptionListener;
-import javax.jms.JMSException;
-
-/**
- * Error listener for listening JMS Connection errors
- */
-public class JMSErrorListener implements ExceptionListener {
-
-    private JMSImprovedConnectionFactory jmsConnectionFactory;
-
-    public JMSErrorListener(JMSImprovedConnectionFactory jmsConnectionFactory) {
-        this.jmsConnectionFactory = jmsConnectionFactory;
-    }
-
-    @Override
-    public void onException(JMSException e) {
-        jmsConnectionFactory.notifyError();
-    }
-}
